requirements.txt
asgiref==3.9.0
async-timeout==5.0.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2.4
djangorestframework==3.16.0
msgpack==1.1.1
redis==6.2.0
sqlparse==0.5.3
tzdata==2025.2
